* Fixes for MAC compilation
git-svn-id: https://svn.code.sf.net/p/klatexformula/code/trunk/klatexformula@413 4112663c-bd99-11de-a44c-fbaffaffe1b7

+++ src/klfbackend/klfdefs.cpp
@@ -148,9 +148,9 @@ KLF_EXPORT QString klfTimeOfDay(bool shortfmt)
   gettimeofday(&tv, NULL);
   char temp[128];
   if (shortfmt)
-    sprintf(temp, "%03ld.%06ld", tv.tv_sec % 1000, tv.tv_usec);
+    sprintf(temp, "%03ld.%06ld", (ulong)tv.tv_sec % 1000, (ulong)tv.tv_usec);
   else
-    sprintf(temp, "%ld.%06ld", tv.tv_sec, tv.tv_usec);
+    sprintf(temp, "%ld.%06ld", (ulong)tv.tv_sec, (ulong)tv.tv_usec);
   return QString::fromAscii(temp);
 }
